refactor to keep cli mjs  clean and facilitate late verification in future
